+++ packages/db/models/subscription/constants.ts
@@ -1,0 +1,2 @@
+export const FREE_BASE_TOKEN_LIMIT = 500000;
+export const FREE_PRO_TOKEN_LIMIT = 25000;