+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @floe/docs
 
+## 0.1.0-beta.2
+
+### Patch Changes
+
+- Updated dependencies [5f84851]
+  - @floe/ui@0.1.0-beta.3
+
 ## 0.1.0-beta.1
 
 ### Minor Changes