+++ packages/db/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @floe/db
 
+## 0.1.0-beta.2
+
+### Minor Changes
+
+- 5f84851: Add support for token usage and pro / basic models.
+
 ## 0.1.0-beta.1
 
 ### Minor Changes